+++ src/main/java/arozxmod/ArozxScreen.java
@@ -1,13 +1,16 @@
 package arozxmod;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.client.option.GameOptions;
 import net.minecraft.screen.ScreenTexts;
 import net.minecraft.text.Text;
 
+@Environment(EnvType.CLIENT)
 public class ArozxScreen extends Screen {
-    private final Screen parent;
+    protected Screen parent;
 
     public ArozxScreen(Screen parent, GameOptions gameOptions) {
         super(Text.literal("Client Mods"));
@@ -44,19 +47,8 @@ public class ArozxScreen extends Screen {
         else
             return (Text.literal("Night vision disabled"));
     }
-    Text autoclickerTextL() {
-        if (ArozxMod.autoclickerEnabledL)
-            return (Text.literal("Left autoclicker enabled"));
-        else
-            return (Text.literal("Left autoclicker disabled"));
-    }
-    Text autoclickerTextR() {
-        if (ArozxMod.autoclickerEnabledL)
-            return (Text.literal("Right autoclicker enabled"));
-        else
-            return (Text.literal("Right autoclicker disabled"));
-    }
 
+    @Override
     protected void init() {
         // Night vision button
         this.addDrawableChild(new ButtonWidget(this.width / 2 - 100, this.height / 6 + 6, 200, 20,
@@ -87,18 +79,6 @@ public class ArozxScreen extends Screen {
                 noFallText(), (button) -> {
             ArozxMod.noFallEnabled = !ArozxMod.noFallEnabled;
             button.setMessage(noFallText());
-        }));
-        // Right autoclicker button
-        this.addDrawableChild(new ButtonWidget(this.width / 2, this.height / 6 + 146, 150, 20,
-                autoclickerTextR(), (button) -> {
-            ArozxMod.autoclickerEnabledR = !ArozxMod.autoclickerEnabledR;
-            button.setMessage(autoclickerTextR());
-        }));
-        // Left autoclicker button
-        this.addDrawableChild(new ButtonWidget(this.width / 2 - 150, this.height / 6 + 146, 150, 20,
-                autoclickerTextL(), (button) -> {
-            ArozxMod.autoclickerEnabledL = !ArozxMod.autoclickerEnabledL;
-            button.setMessage(autoclickerTextL());
         }));
         // Back button
         this.addDrawableChild(new ButtonWidget(this.width / 2 - 100, this.height / 6 + 174, 200, 20,