+++ src/main/java/arozxmod/mixin/GameMenuScreenMixin.java
@@ -1,6 +1,7 @@
 package arozxmod.mixin;
 
 import arozxmod.ArozxScreen;
+import arozxmod.AutoclickerScreen;
 import net.minecraft.client.gui.screen.GameMenuScreen;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.ButtonWidget;
@@ -19,6 +20,10 @@ public abstract class GameMenuScreenMixin extends Screen {
         this.addDrawableChild(new ButtonWidget(10, 10, 90, 20, Text.literal("Mod Menu"), (button) -> {
 
             this.client.setScreen(new ArozxScreen(this, this.client.options));
+        }));
+        this.addDrawableChild(new ButtonWidget(10, 30, 90, 20, Text.literal("Autoclicker"), (button) -> {
+
+            this.client.setScreen(new AutoclickerScreen(this, this.client.options));
         }));
     }
 }