+++ Changelog.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+
+### Added
+
+- New menu
+
+### Changed
+
+- Moved autoclicker to new menu
+- Updated fabric.mod.json
+- Changed classtypes in ArozxScreen.java
+
+
 ## [1.0.1] - 2023-07-05
 
 ### Added